refactor(controllers): collapse the SettingsModelSelectionControllerFactory wrapper
Closes #10620
